apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "prospectiq-api"
version = "0.1.0"
description = "FastAPI backend foundation for LeadScope AI"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.13.2",
  "email-validator>=2.2.0",
  "fastapi>=0.115.0",
  "httpx>=0.27.0",
  "pydantic-settings>=2.3.4",
  "PyJWT>=2.9.0",
  "PyMySQL>=1.1.1",
  "pwdlib[argon2]>=0.2.0",
  "SQLAlchemy>=2.0.36",
  "uvicorn[standard]>=0.30.6",
]

[project.optional-dependencies]
dev = [
  "mypy>=1.11.1",
  "pre-commit>=3.8.0",
  "pytest>=8.3.2",
  "pytest-asyncio>=0.24.0",
  "ruff>=0.6.4",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py312"
extend-exclude = ["migrations"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "A"]
ignore = ["B008", "E501"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.12"
strict = true
packages = ["app"]
exclude = ["migrations"]
warn_unused_configs = true
